% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_path}
\alias{check_path}
\title{Check github path}
\usage{
check_path(.path)
}
\arguments{
\item{.path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}
}
\value{
The GitHub repo path
}
\description{
Check github path
}
\keyword{internal}
