% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_check_assignable}
\alias{gh_check_assignable}
\title{Check that a person can be assigned to an issue}
\usage{
gh_check_assignable(path, assignee, ..., messages = TRUE)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
TRUE/FALSE if a collaborator is assignable
}
\description{
Check that a person can be assigned to an issue
}
\keyword{internal}
