% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_invite}
\alias{gh_collab_invite}
\title{Invite a collaborator to a Github repo}
\usage{
gh_collab_invite(path, invitee, ...)
}
\arguments{
\item{path}{Github repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{invitee}{The person to invite.}

\item{...}{Additional parameters, see https://docs.github.com/rest/reference/repos#add-a-repository-collaborator}
}
\description{
Invite a collaborator to a Github repo
}
