% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_list}
\alias{gh_collab_list}
\title{List current collaborators for a GitHub repo}
\usage{
gh_collab_list(path, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
A data frame that contains the collaborators for the specified repository.
}
\description{
List current collaborators for a GitHub repo
}
\seealso{
\code{\link[=gh_collab_check]{gh_collab_check()}} for checking whether a specific user is a collaborator.

Other Collaborations: 
\code{\link{gh_collab_check}()},
\code{\link{gh_collab_invite}()},
\code{\link{gh_collab_pending}()},
\code{\link{gh_collab_remove}()},
\code{\link{gh_collab_uninvite}()}
}
\concept{Collaborations}
