% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_pending}
\alias{gh_collab_pending}
\title{List pending invitations for a repo}
\usage{
gh_collab_pending(path, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
A dataframe listing the usernames of pending invitees (or dataframe of length 0 if no pending invitees).
}
\description{
List pending invitations for a repo
}
\seealso{
Other Collaborations: 
\code{\link{gh_collab_check}()},
\code{\link{gh_collab_invite}()},
\code{\link{gh_collab_list}()},
\code{\link{gh_collab_remove}()},
\code{\link{gh_collab_uninvite}()}
}
\concept{Collaborations}
