% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_remove}
\alias{gh_collab_remove}
\title{Remove a collaborator from a repository}
\usage{
gh_collab_remove(path, collaborator, ..., messages = TRUE)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{collaborator}{The collaborator for whom to call this function.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}

\item{messages}{Logical: Should a message indicating the status of the function be printed? TRUE by default.}
}
\value{
NA; called for its side effects.
}
\description{
This will remove a collaborator from a repository and confirm the individual is no longer a member.
}
