% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_uninvite}
\alias{gh_collab_uninvite}
\title{Uninvite a collaborator to a GitHub repo}
\usage{
gh_collab_uninvite(path, collaborator, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{collaborator}{The collaborator for whom to call this function.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
NA; called for side effects to uninvite a collaborator to the specified GitHub repository.
}
\description{
This only works if the individual's invite is pending.
}
\seealso{
Other Collaborations: 
\code{\link{gh_collab_check}()},
\code{\link{gh_collab_invite}()},
\code{\link{gh_collab_list}()},
\code{\link{gh_collab_pending}()},
\code{\link{gh_collab_remove}()}
}
\concept{Collaborations}
