% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_common.R
\name{gh_issue}
\alias{gh_issue}
\title{Create an issue on a github repo}
\usage{
gh_issue(path, title, body, ...)
}
\arguments{
\item{path}{Github repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{title}{The issue's title}

\item{body}{The issue's body text}

\item{...}{Additional parameters, see https://docs.github.com/en/rest/reference/issues#create-an-issue}
}
\description{
Create an issue on a github repo
}
