% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_comment}
\alias{gh_issue_comment}
\title{Comment on a GitHub issue}
\usage{
gh_issue_comment(path, issue_number, body, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{issue_number}{The issue number on which to comment. Can be determined online or with \code{\link{gh_issue_list}}}

\item{body}{The issue's body text}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
NA; called for side effects.
}
\description{
Comment on a GitHub issue
}
\seealso{
Other Issues: 
\code{\link{gh_issue_assign}()},
\code{\link{gh_issue_list}()},
\code{\link{gh_issue_mention}()},
\code{\link{gh_issue_new}()}
}
\concept{Issues}
