% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_list}
\alias{gh_issue_list}
\title{List issues for a GitHub Repo}
\usage{
gh_issue_list(path, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
A list of issues for the specified GitHub repo
}
\description{
List issues for a GitHub Repo
}
