% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_mention}
\alias{gh_issue_mention}
\title{Check to see if a person was mentioned in a GitHub issue}
\usage{
gh_issue_mention(path, collaborator, ...)
}
\arguments{
\item{...}{}
}
\value{
TRUE if the collaborator was mentioned in any issues; FALSE otherwise
}
\description{
Check to see if a person was mentioned in a GitHub issue
}
