% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repos.R
\name{gh_repos_create}
\alias{gh_repos_create}
\title{Create a new repository}
\usage{
gh_repos_create(path, private = TRUE, description = NULL, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{private}{Will this repository be private? TRUE/FALSE; default is TRUE.}

\item{description}{(Optional) A short description of the repository.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\description{
Create a new repository
}
