% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repos.R
\name{gh_repos_delete_internal}
\alias{gh_repos_delete_internal}
\title{Actually delete repo}
\usage{
gh_repos_delete_internal(owner, repo, ...)
}
\arguments{
\item{owner}{GitHub user}

\item{repo}{GitHub repo}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
NA; used for side effects to delete a repository from GitHub.
}
\description{
Actually delete repo
}
\seealso{
Other Repositories: 
\code{\link{gh_repos_create}()},
\code{\link{gh_repos_delete}()},
\code{\link{gh_repos_list}()},
\code{\link{gh_repos_mutate}()}
}
\concept{Repositories}
\keyword{internal}
