% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_repos.R
\name{gh_repos_list}
\alias{gh_repos_list}
\title{List a user's github repository}
\usage{
gh_repos_list(user, ...)
}
\arguments{
\item{user}{The GitHub user}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
Repositories for the specified user
}
\description{
List a user's github repository
}
\examples{
gh_repos_list("jdtrat")
}
