% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_check_assignable}
\alias{gh_check_assignable}
\title{Check that a person can be assigned to an issue}
\usage{
gh_check_assignable(path, assignee, messages = TRUE)
}
\arguments{
\item{assignee}{}
}
\description{
Check that a person can be assigned to an issue
}
\keyword{internal}
