% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_collab.R
\name{gh_collab_check}
\alias{gh_collab_check}
\title{Check if a user is a collaborator}
\usage{
gh_collab_check(path, collaborator, messages = TRUE)
}
\arguments{
\item{messages}{}
}
\value{

}
\description{
Check if a user is a collaborator
}
