% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_assign}
\alias{gh_issue_assign}
\title{Assign people to a Github issue}
\usage{
gh_issue_assign(path, issue_number, assignees)
}
\arguments{
\item{assignees}{}
}
\description{
Assign people to a Github issue
}
