% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_comment}
\alias{gh_issue_comment}
\title{Comment on a Github issue}
\usage{
gh_issue_comment(path, issue_number, body, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Comment on a Github issue
}
