% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_list}
\alias{gh_issue_list}
\title{List issues for a Github Repo}
\usage{
gh_issue_list(path, ...)
}
\arguments{
\item{path}{Github repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{...}{Additional parameters, see https://docs.github.com/en/rest/reference/issues#list-repository-issues}
}
\value{

}
\description{
List issues for a Github Repo
}
