% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_mention}
\alias{gh_issue_mention}
\title{Check to see if a person was mentioned in a GitHub issue}
\usage{
gh_issue_mention(path, collaborator, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{collaborator}{The collaborator for whom to call this function.}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
TRUE if the collaborator was mentioned in any issues; FALSE otherwise
}
\description{
Check to see if a person was mentioned in a GitHub issue
}
\seealso{
Other Issues: 
\code{\link{gh_issue_assign}()},
\code{\link{gh_issue_comment}()},
\code{\link{gh_issue_list}()},
\code{\link{gh_issue_new}()}
}
\concept{Issues}
