% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_new}
\alias{gh_issue_new}
\title{Create an issue on a github repo}
\usage{
gh_issue_new(path, title, body, ...)
}
\arguments{
\item{path}{GitHub repo path of the form "{username}/{repo}". If not provided, will default to current repo.}

\item{title}{The issue's title}

\item{body}{The issue's body text}

\item{...}{Additional parameters, see \url{https://docs.github.com/en/rest}}
}
\value{
NA; used for side effects to create a new issue on the specified GitHub repository.
}
\description{
Create an issue on a github repo
}
\seealso{
Other Issues: 
\code{\link{gh_issue_assign}()},
\code{\link{gh_issue_comment}()},
\code{\link{gh_issue_list}()},
\code{\link{gh_issue_mention}()}
}
\concept{Issues}
